Factor out ChannelRequester, report connection status
