+++ src/EpicsClient/EpicsClientRandom.h
@@ -28,6 +28,8 @@ public:
   /// Generate a fake EpicsPVUpdate and emit it
   void generateFakePVUpdate();
 
+  std::string getConnectionStatus() override { return "EpicsClientRandom"; }
+
 private:
   /// Get current time since unix epoch in nanoseconds
   uint64_t getCurrentTimestamp() const;
